@@ -1,11 +1,4 @@
-const path = require('path');
-module.exports = {
-    outputDir: path.resolve(__dirname, '../src/main/resources/static'),
-    indexPath: '../templates/home.html',
-    configureWebpack : {
-        name: 'test',
-        output: {
-
-        }
-    }
-}
+const { defineConfig } = require('@vue/cli-service')
+module.exports = defineConfig({
+  transpileDependencies: true
+})
